backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
pdfplumber==0.10.3
python-docx==1.1.0
pandas==2.1.3
openpyxl==3.1.2
pytesseract==0.3.10
Pillow==10.1.0
sentence-transformers>=2.3.0
openai==1.3.5
reportlab==4.0.7
numpy==1.26.2
scikit-learn==1.3.2
langchain>=1.0.0
langchain-community>=0.4.0
langchain-text-splitters>=1.0.0
faiss-cpu>=1.7.4
pypdf>=4.0.0